rename methods and packages to match refactoring
