+++ archivist/src/test/scala/uk/ac/wellcome/platform/archive/archivist/flow/UploadDigestItemFlowTest.scala
@@ -44,7 +44,7 @@ class UploadDigestItemFlowTest
           val digest =
             "52dbe81fda7f771f83ed4afc9a7c156d3bf486f8d654970fa5c5dbebb4ff7b73"
 
-          val archiveItemJob = createArchiveDigestItemJobWith(
+          val archiveItemJob = createDigestItemJobWith(
             file = file,
             bucket = bucket,
             digest = digest,
@@ -81,7 +81,7 @@ class UploadDigestItemFlowTest
         withZipFile(List(FileEntry(s"$fileName", fileContent))) { file =>
           val digest = "wrong!"
 
-          val archiveItemJob = createArchiveDigestItemJobWith(
+          val archiveItemJob = createDigestItemJobWith(
             file = file,
             bucket = bucket,
             digest = digest,
@@ -114,7 +114,7 @@ class UploadDigestItemFlowTest
           val fileName = "key.txt"
           val bagIdentifier = createExternalIdentifier
 
-          val archiveItemJob = createArchiveDigestItemJobWith(
+          val archiveItemJob = createDigestItemJobWith(
             file = file,
             bucket = bucket,
             bagIdentifier = bagIdentifier,
@@ -148,7 +148,7 @@ class UploadDigestItemFlowTest
 
       val fileName = "key.txt"
       withZipFile(List(FileEntry(s"$fileName", fileContent))) { file =>
-        val failingArchiveItemJob = createArchiveDigestItemJobWith(
+        val failingArchiveItemJob = createDigestItemJobWith(
           file = file,
           bucket = Bucket("does-not-exist"),
           itemPath = fileName
